requirements.txt
# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
aiofiles>=23.2.1

# HTTP Client
requests>=2.31.0
httpx>=0.27.0
aiohttp>=3.9.0

# Data Validation
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Flask (للخدمات القديمة)
flask>=3.0.0
flask-cors>=4.0.0

# AI/ML - PyTorch
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers & NLP
transformers>=4.30.0
tokenizers>=0.15.0
sentence-transformers>=2.2.0
accelerate>=0.25.0

# Vector Database
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
aiosqlite>=0.19.0
asyncpg>=0.29.0

# Caching & Queue
redis>=5.0.0
celery>=5.3.0

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=23.2.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.27.0

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
pre-commit>=3.6.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0.1
python-jose[cryptography]>=3.3.0

# API Documentation
fastapi-pagination>=0.12.0
